Adding .sh script.

+++ file_out.py
@@ -1,4 +1,4 @@
-import subprocess
+import os
 from datetime import datetime
 
 import pandas as pd
@@ -24,18 +24,25 @@ class SimulationLog:
     def get_file(self) -> {}:
         return self._data
 
-    def save_file(self) -> None:
+    def save_file(self, alg: str) -> None:
         """Save log file with current date and time."""
         today = datetime.now()
-        file_name = str("./output/SIMULATION_%s.csv" % today.strftime("%d_%m_%Y_%H%M%S"))
+        file_name = str("./output/%s_SIMULATIONS_%s.csv" % (alg, today.strftime("%d%m%Y")))
         df = pd.DataFrame.from_dict(self._data, "index")
-        # print(self._file.head(10))
         try:
-            output_dir = subprocess.check_call(["mkdir", "-p", "output"])
-            f = open(file_name, "w")
-            df.to_csv(f, index=False, na_rep=0)
-        except subprocess.CalledProcessError:
-            f = open("SIMULATION_ % s.csv" % today.strftime("%d_%m_%Y_%H%M%S"), "w")
-            df.to_csv(f, index=False, na_rep=0)
+            os.mkdir('./output')
+        except FileExistsError:
+            pass
+        try:
+            if os.path.exists(file_name):
+                f = open(file_name, "a")
+                df.to_csv(f, index=False, na_rep=0, header=False)
+            else:
+                f = open(file_name, "w")
+                df.to_csv(f, index=False, na_rep=0)
+        except IOError:
+            print('IO Error: File did not save.')
+        finally:
+            f.close()
 
         print('File Printed to %s' % file_name)
